decompose ticket timeline item into subcomponents
