+++ app/src/features/admin/tickets/hooks/useTicketComment.ts
@@ -1,15 +1,11 @@
-import {
-  TICKET_ATTACHMENT_IMAGE_FILE_EXTENSIONS,
-  TICKET_ATTACHMENT_MARKDOWN_FORMATTERS,
-  TICKET_ATTACHMENT_MARKDOWN_TYPE_BY_MEDIA_TYPE,
-  TicketAttachmentMarkdownType
-} from 'constants/ticket';
+import { getArtifactMarkdownByMimeType } from 'features/admin/tickets/utils/ticketArtifactMarkdown';
 import { APIError } from 'hooks/api/useAxios';
 import { useApi } from 'hooks/useApi';
 import { useAuthStateContext } from 'hooks/useAuthStateContext';
-import { useConfigContext, useDialogContext, useTicketContext } from 'hooks/useContext';
-import { ITicketArtifact, ITicketCommentLog } from 'interfaces/useTicketsApi.interface';
+import { useDialogContext, useTicketContext } from 'hooks/useContext';
+import { ITicketCommentLog } from 'interfaces/useTicketsApi.interface';
 import { useState } from 'react';
+import { useTicketAttachmentUpload } from './useTicketAttachmentUpload';
 
 /**
  * Comment state and submit behavior for ticket details.
@@ -19,60 +15,12 @@ import { useState } from 'react';
 export const useTicketComment = () => {
   const api = useApi();
   const { ticketId, ticketDataLoader } = useTicketContext();
-  const config = useConfigContext();
   const authStateContext = useAuthStateContext();
   const dialogContext = useDialogContext();
+  const { isUploadingAttachment, uploadTicketAttachment } = useTicketAttachmentUpload({ ticketId });
 
   const [comment, setComment] = useState('');
   const [isSavingComment, setIsSavingComment] = useState(false);
-  const [isUploadingAttachment, setIsUploadingAttachment] = useState(false);
-
-  /**
-   * Determine the markdown rendering type for an uploaded attachment.
-   *
-   * Browser `File.type` values are MIME types such as `image/png`, so the first
-   * segment maps image MIME types to markdown image syntax. Some browsers or
-   * operating systems may provide an empty MIME type, so image file extensions
-   * are used as a fallback before defaulting to a standard markdown link.
-   *
-   * @param {File} file File selected by the user.
-   * @returns {TicketAttachmentMarkdownType} Markdown formatter type for the file.
-   */
-  const getAttachmentMarkdownType = (file: File): TicketAttachmentMarkdownType => {
-    const mediaType = file.type.split('/')[0]?.toLowerCase();
-    const markdownType = mediaType ? TICKET_ATTACHMENT_MARKDOWN_TYPE_BY_MEDIA_TYPE[mediaType] : undefined;
-
-    if (markdownType) {
-      return markdownType;
-    }
-
-    const extension = file.name.split('.').pop()?.toLowerCase();
-
-    return extension && TICKET_ATTACHMENT_IMAGE_FILE_EXTENSIONS.has(extension) ? 'image' : 'link';
-  };
-
-  /**
-   * Convert an uploaded ticket attachment into the markdown syntax inserted into
-   * the comment draft.
-   *
-   * The file's markdown type is resolved from its MIME type, with a file
-   * extension fallback when MIME metadata is unavailable, and then formatted
-   * through `TICKET_ATTACHMENT_MARKDOWN_FORMATTERS`. Image files are inserted as
-   * markdown images so the preview renders them as attachment image references.
-   * Other files are inserted as standard markdown links. Both variants use the
-   * stable `ticket_artifact_id` returned after the upload is completed.
-   *
-   * @param {File} file File selected by the user.
-   * @param {string} ticketArtifactId Stable ticket artifact identifier.
-   * @returns {string} Markdown image or link text for the uploaded attachment.
-   */
-  const getArtifactMarkdownByMimeType = (file: File, ticketArtifactId: string) => {
-    const label = file.name;
-    const href = `/artifact/${ticketArtifactId}`;
-    const markdownType = getAttachmentMarkdownType(file);
-
-    return TICKET_ATTACHMENT_MARKDOWN_FORMATTERS[markdownType](label, href);
-  };
 
   /**
    * Append a newly created comment to the locally cached ticket details.
@@ -240,73 +188,22 @@ export const useTicketComment = () => {
    * @returns {Promise<void>} Resolves when the upload attempt has completed.
    */
   const handleUploadAttachment = async (file: File) => {
-    const maxTicketAttachmentFileSize = config.MAX_TICKET_ATTACHMENT_FILE_SIZE;
+    const ticketArtifact = await uploadTicketAttachment(file);
 
-    if (file.size > maxTicketAttachmentFileSize) {
-      const maxTicketAttachmentFileSizeMB = Math.round(maxTicketAttachmentFileSize / 1024 / 1024);
-
-      dialogContext.setSnackbar({
-        open: true,
-        snackbarMessage: `Attachment exceeds the ${maxTicketAttachmentFileSizeMB} MB limit.`
-      });
+    if (!ticketArtifact) {
       return;
     }
 
-    try {
-      setIsUploadingAttachment(true);
+    const markdownLink = getArtifactMarkdownByMimeType(file, ticketArtifact.ticket_artifact_id);
 
-      const initializedUpload = await api.tickets.createTicketUpload(ticketId, {
-        file_name: file.name,
-        byte_size: file.size,
-        content_type: file.type || 'application/octet-stream'
-      });
-
-      const uploadResponse = await fetch(initializedUpload.presigned_upload_url, {
-        method: 'PUT',
-        headers: {
-          'Content-Type': file.type || 'application/octet-stream'
-        },
-        body: file
-      });
-
-      if (!uploadResponse.ok) {
-        throw new Error('Failed to upload attachment.');
+    setComment((previousComment) => {
+      if (!previousComment) {
+        return markdownLink;
       }
 
-      const ticketArtifact = await api.tickets.completeTicketUpload(ticketId, initializedUpload.upload_id, {
-        status: 'uploaded'
-      });
-      const markdownLink = getArtifactMarkdownByMimeType(file, ticketArtifact.ticket_artifact_id);
-
-      const latestTicket = ticketDataLoader.data;
-      if (latestTicket) {
-        ticketDataLoader.setData({
-          ...latestTicket,
-          artifacts: latestTicket.artifacts.some(
-            (artifact: ITicketArtifact) => artifact.ticket_artifact_id === ticketArtifact.ticket_artifact_id
-          )
-            ? latestTicket.artifacts
-            : [...latestTicket.artifacts, ticketArtifact]
-        });
-      }
-
-      setComment((previousComment) => {
-        if (!previousComment) {
-          return markdownLink;
-        }
-
-        const separator = /\s$/.test(previousComment) ? '' : ' ';
-        return `${previousComment}${separator}${markdownLink}`;
-      });
-    } catch (caughtError) {
-      const apiError = caughtError as APIError;
-      dialogContext.setSnackbar({
-        open: true,
-        snackbarMessage: apiError.message || 'Failed to upload attachment.'
-      });
-    } finally {
-      setIsUploadingAttachment(false);
-    }
+      const separator = /\s$/.test(previousComment) ? '' : ' ';
+      return `${previousComment}${separator}${markdownLink}`;
+    });
   };
 
   return {
